Set up basic skeleton for API routes and db queries
