+++ ProductsAPI/MySQL/controllers/styles.js
@@ -1,0 +1,7 @@
+const models = require('../models');
+
+const get = (req, res) => {
+
+};
+
+module.exports = get;
