+++ ProductsAPI/MySQL/database/index.js
@@ -7,4 +7,6 @@ const db = mysql.createConnection({
   database: 'productsSDC'
 });
 
+db.connect();
+
 module.exports = db;
