+++ ProductsAPI/MySQL/models/index.js
@@ -1,0 +1,4 @@
+exports.products = require('./products.js');
+exports.product_id = require('./product_id.js');
+exports.styles = require('./styles.js');
+exports.related = require('./related.js');
