+++ ProductsAPI/MySQL/models/product_id.js
@@ -1,0 +1,7 @@
+const db = require('../database');
+
+const getAll = (callback) => {
+
+};
+
+module.exports = getAll;
