+++ ProductsAPI/MySQL/models/products.js
@@ -1,0 +1,7 @@
+const db = require('../database');
+
+const getAll = (callback) => {
+
+};
+
+module.exports = getAll;
